moving seasons, seasonslist and showsearch
