+++ src/containers/ShowSearch/ShowSearch.scss
@@ -1,0 +1,8 @@
+.ShowSearch {
+  z-index: 1000;
+  position: absolute;
+  top: 11px;
+  aside {
+    background-color: #375a7f;
+  }
+}
